Update checkout config
Update checkout config
