+++ Sources/VGSCheckoutSDK/Core/Checkout/Models/CheckoutRequestResult/VGSCheckoutRequestResult.swift
@@ -4,6 +4,12 @@
 
 import Foundation
 
+/// Checkout addtional info in `VGSCheckoutRequestResult`.
+public protocol VGSCheckoutInfo {}
+
+/// Basic Checkout addtional info.
+internal protocol VGSCheckoutBasicExtraData {}
+
 /// Response enum cases for SDK requests.
 @frozen public enum VGSCheckoutRequestResult {
 		/**
@@ -13,8 +19,9 @@ import Foundation
 				- code: response status code.
 				- data: response **data** object.
 				- response: URLResponse object represents a URL load response.
+				- info: `VGSCheckoutInfo?`, addtional info in `VGSCheckoutRequestResult`.
 		*/
-		case success(_ code: Int, _ data: Data?, _ response: URLResponse?)
+		case success(_ code: Int, _ data: Data?, _ response: URLResponse?, _ info: VGSCheckoutInfo? = nil)
 
 		/**
 		 Failed response case
@@ -24,6 +31,7 @@ import Foundation
 				- data: response **Data** object.
 				- response: `URLResponse` object represents a URL load response.
 				- error: `Error` object.
+				- info: `VGSCheckoutInfo?`, addtional info in `VGSCheckoutRequestResult`.
 		*/
-		case failure(_ code: Int, _ data: Data?, _ response: URLResponse?, _ error: Error?)
+		case failure(_ code: Int, _ data: Data?, _ response: URLResponse?, _ error: Error?, _ info: VGSCheckoutInfo? = nil)
 }
