+++ Sources/VGSCheckoutSDK/Core/Collect/CollectCore/VGSError.swift
@@ -26,6 +26,9 @@ internal enum VGSErrorType: Int {
 
 		/// When VGS config URL is not valid.
 		case invalidConfigurationURL = 1480
+
+		/// When multiplexing JWT token is not valid.
+		case invalidJWTToken = 1484
 }
 
 /// An error produced by `VGSCheckout`. Works similar to default `NSError` in iOS.
@@ -53,3 +56,6 @@ internal class VGSError: NSError {
         super.init(domain: VGSCheckoutErrorDomain, code: type.rawValue, userInfo: info?.asDictionary)
     }
 }
+
+
+
