+++ Sources/VGSCheckoutSDK/Core/Configuration/Multiplexing/VGSCheckoutMultiplexingAddCardConfiguration.swift
@@ -1,12 +1,22 @@
 //
-//  VGSCheckoutSDKMultiplexingConfiguration.swift
+//  VGSCheckoutMultiplexingAddCardConfiguration.swift
 //  VGSCheckoutSDK
 
 import Foundation
 
 /// Holds configuration with predefined setup for work with payment orchestration/multiplexing app, confirms to `VGSCheckoutBasicConfigurationProtocol`.
-public struct VGSCheckoutMultiplexingConfiguration: VGSCheckoutBasicConfigurationProtocol {
-  
+public struct VGSCheckoutMultiplexingAddCardConfiguration: VGSCheckoutBasicConfigurationProtocol {
+
+	/// A callback to be run with a `VGSCheckoutMultiplexingAddCardConfiguration` on configuration setup succeed.
+	/// - Parameters:
+	///   - configuration:  `VGSCheckoutMultiplexingAddCardConfiguration` object, configuration. 
+	public typealias CreateConfigurationSuccessCompletion = (_ configuration: inout VGSCheckoutMultiplexingAddCardConfiguration) -> Void
+
+	/// A callback to be run with an error when configuration setup fail.
+	/// - Parameters:
+	///   - error: `Error` object, the error on configuration setup fail.
+	public typealias CreateConfigurationFailCompletion = (_ error: Error) -> Void
+
   // MARK: - Attributes
   
 	/// `String` object, payment orchestration tenant id.
@@ -18,17 +28,34 @@ public struct VGSCheckoutMultiplexingConfiguration: VGSCheckoutBasicConfiguratio
 	/// Payment orchestration access token.
 	private(set) public var accessToken: String
 
+	// MARK: - Public
+
+	/// Creates Multiplexing Save Card config
+	/// - Parameters:
+	///   - accessToken: `String` object, should be valid access token for payment orchestration.
+	///   - tenantId: `String` object, payment orchestration tenant id.
+	///   - environment: `String` object, organization vault environment with data region.(e.g. "live", "live-eu1", "sandbox"). Default is `sandbox`.
+	///   - success: `CreateConfigurationSuccessCompletion` object, callback for configuration setup succeed.
+	///   - failure: `CreateConfigurationFailCompletion` object, callback for configuration setup fail.
+	public static func createConfiguration(accessToken: String, tenantId: String, environment: String = "sandbox", success: @escaping CreateConfigurationSuccessCompletion, failure: @escaping CreateConfigurationFailCompletion) {
+		guard VGSMultiplexingCredentialsValidator.isJWTScopeValid(accessToken, vaultId: tenantId, environment: environment) else {
+			let error = NSError(domain: VGSCheckoutErrorDomain, code: VGSErrorType.invalidJWTToken.rawValue, userInfo: [NSLocalizedDescriptionKey: "JWT token is invalid or empty!"])
+			failure(error as Error)
+			return
+		}
+
+		var saveCardConfiguration = VGSCheckoutMultiplexingAddCardConfiguration(accessToken: accessToken, tenantId: tenantId, environment: environment)
+		success(&saveCardConfiguration)
+	}
+
   // MARK: - Initialization
   
-	/// Configuration initializer (failable).
+	/// Configuration initializer.
 	/// - Parameters:
   ///   - accessToken: `String` object, should be valid access token for payment orchestration.
 	///   - tenantId: `String` object, payment orchestration tenant id.
 	///   - environment: `String` object, organization vault environment with data region.(e.g. "live", "live-eu1", "sandbox"). Default is `sandbox`.
-	public init?(accessToken: String, tenantId: String, environment: String = "sandbox") {
-		guard VGSMultiplexingCredentialsValidator.isJWTScopeValid(accessToken, vaultId: tenantId, environment: environment) else {
-            return nil
-        }
+	internal init(accessToken: String, tenantId: String, environment: String = "sandbox") {
     self.accessToken = accessToken
 		self.tenantId = tenantId
 		self.environment = environment
